Solucionando errores
- Fallaba al obtener SSID a partir de Android 12
- Solo mostraba el primer hogar del usuario
- Crasheaba con datos de medición

+++ app/src/main/java/es/uma/smartmeter/MedicionFragment.java
@@ -43,7 +43,7 @@ public class MedicionFragment extends Fragment {
                 binding.include.tvPower.setText(jsonObject.get("kw") + "KW");
                 int progress = 0;
                 int max_kw = 3000; //KW máximos alcanzables estimados. Usado para calcular el porcentage de progressbar coloreado
-                progress += Integer.parseInt(jsonObject.get("kw").toString());
+                progress += Double.parseDouble(jsonObject.get("kw").toString());
                 progress = progress/max_kw * 100;
                 binding.include.progressBar.setProgress(progress);
                 System.out.println(binding.include.tvPower.getText());
